Removed the need to use the variable "orderMaximum"
Previously named "ORDMAX", this variable controlled the size of the array which stores odd multiples. ORDMAX had to be user-inputted, which made little sense because the user may not know or understand the variable's use. It has been replaced by numberOfPrimes/10. This way, this variable grows as the number of requested primes grows, which is useful because the array it controls must indeed be larger in this case.